добавлен 4 вариант в Task2

+++ Урок_5/Lesson5_Task2_3/Lesson5_Task2_3/Program.cs
@@ -7,8 +7,8 @@ namespace Lesson5_Task2_3
         static void Main(string[] args)
         {
             Console.WriteLine("Введите число");
-            byte a = Convert.ToByte(Console.ReadLine());
-            byte result = (byte)(a << 7);
+            int a = Convert.ToInt32(Console.ReadLine());
+            int result = (byte)(a << 7);
 
             if (result==0)
             {
